+++ functions.php
@@ -1,6 +1,11 @@
 <?php
 
 /**
+ * Theme Setup.
+ */
+require get_theme_file_path( '/inc/setup.php' );
+
+/**
  * Load Enqueue Scripts.
  */
 require get_theme_file_path( '/inc/enqueue.php' );