+++ inc/setup.php
@@ -1,0 +1,53 @@
+<?php
+
+if ( ! function_exists( 'civ_setup' ) ) :
+  /**
+   * Sets up theme defaults and registers support for various WordPress features.
+   */
+  function civ_setup() {
+    // Add default posts and comments RSS feed links to head.
+    add_theme_support( 'automatic-feed-links' );
+
+    // Let WordPress manage the document title.
+    add_theme_support( 'title-tag' );
+
+    // Enable support for Post Thumbnails on posts and pages.
+    add_theme_support( 'post-thumbnails' );
+
+    // This theme uses wp_nav_menu() in one location.
+    register_nav_menus(
+      array(
+        'menu-1' => esc_html__( 'Primary', 'civ-shows' ),
+      )
+    );
+
+    /*
+     * Switch default core markup for search form, comment form, and comments
+     * to output valid HTML5.
+     */
+    add_theme_support(
+      'html5',
+      array(
+        'search-form',
+        'comment-form',
+        'comment-list',
+        'gallery',
+        'caption',
+        'style',
+        'script',
+      )
+    );
+
+    // Custom logo support.
+    add_theme_support(
+      'custom-logo',
+      array(
+        'height'      => 250,
+        'width'       => 250,
+        'flex-width'  => true,
+        'flex-height' => true,
+      )
+    );
+  }
+endif;
+add_action( 'after_setup_theme', 'civ_setup' );